+++ src/componant/Nav/NavLogo.js
@@ -1,4 +1,4 @@
-import logo from '../images/logo.png';
+import logo from '../../images/logo.png';
 
 function NavLogo({ className }) {
   return (
