+++ src/index.js
@@ -3,6 +3,7 @@ import ReactDOM from 'react-dom/client';
 import App from './App';
 import './styles/reset.css';
 import './styles/index.css';
+import './fonts/pretendard/pretendard.css';
 
 const root = ReactDOM.createRoot(document.getElementById('root'));
 root.render(<App />);
